패키지 구조 변경 (객체 지향적) 및 디펜던시 추가
