--- src/main/java/site/gaeddocoding/deliveryproject/domain/food/Food.java
@@ -1,12 +1,0 @@
-package site.gaeddocoding.deliveryproject.domain.food;
-
-
-public class Food {
-
-    private Integer id;
-    private String foodName;
-    private Integer foodPrice;
-    private Integer foodAmt;
-    private String foodSize;
-    private boolean sideDish;
-}
