+++ src/main/java/site/gaeddocoding/deliveryproject/food/FoodController.java
@@ -1,0 +1,4 @@
+package site.gaeddocoding.deliveryproject.food;
+
+public class FoodController {
+}
