+++ src/main/java/site/gaeddocoding/deliveryproject/food/FoodRepository.java
@@ -1,0 +1,4 @@
+package site.gaeddocoding.deliveryproject.food;
+
+public interface FoodRepository {
+}
