--- src/main/java/site/gaeddocoding/deliveryproject/service/FoodService.java
@@ -1,4 +1,0 @@
-package site.gaeddocoding.deliveryproject.service;
-
-public class FoodService {
-}
