--- src/main/java/site/gaeddocoding/deliveryproject/web/FoodController.java
@@ -1,4 +1,0 @@
-package site.gaeddocoding.deliveryproject.web;
-
-public class FoodController {
-}
