--- src/main/java/site/gaeddocoding/deliveryproject/web/dto/food/FoodDto.java
@@ -1,4 +1,0 @@
-package site.gaeddocoding.deliveryproject.web.dto.food;
-
-public class FoodDto {
-}
